Fix test flakyness in generate_cert_test.go
The flakyness was due to we set the NotBefore field to current time.
However when verifying the cert, the NotBefore field is required to be
strictly earlier than the time at which verification happens. Thus
test fails occasionally if there isn't sufficient time elapsed between
cert issuance and cert verification.

Fixes #253